@@ -8,58 +8,70 @@
   "bin": "./index.js",
   "scripts": {
     "start": "webpack-serve webpack.config.dev.js",
+    "start:prod": "webpack-serve webpack.config.prod.js",
     "build": "yarn build:dev",
     "build:dev": "webpack --mode=development --config webpack.config.dev.js",
     "build:prod": "webpack --mode=production --config webpack.config.prod.js",
     "lint": "tslint -p ./ -c ./tslint.json './src/**/*.ts{,x}'",
     "test": "yarn lint && yarn build",
+    "size-analyze": "yarn build:prod --json | tail -n +2 | webpack-bundle-size-analyzer",
     "docs": "typedoc --excludeExternals --externalPattern '**/node_modules/**' --ignoreCompilerErrors --out ./docs/ ./src/ && echo 'include:\n  - \"_*_.html\"\n  - \"_*_.*.html\"' > docs/_config.yml",
     "deploy": "yarn lint && yarn build:prod && aws s3 sync ./assets s3://souzou-portal/ --exclude '.DS_Store' --acl public-read --cache-control 'max-age=3600' && aws s3 sync ./dist s3://souzou-portal/  --exclude '.DS_Store' --acl public-read --cache-control 'max-age=3600'"
   },
   "dependencies": {
-    "@material-ui/core": "^1.3.1",
-    "@material-ui/icons": "^1.1.0",
-    "amazon-cognito-identity-js": "^2.0.13",
-    "apollo-client": "^2.3.5",
-    "aws-appsync": "^1.2.1",
-    "aws-appsync-react": "^1.0.21",
-    "graphql": "^0.13.2",
+    "@material-ui/core": "^3.0.1",
+    "@material-ui/icons": "^3.0.1",
+    "amazon-cognito-identity-js": "^2.0.23",
+    "apollo-client": "^2.4.1",
+    "aws-appsync": "^1.3.4",
+    "aws-appsync-react": "^1.1.4",
+    "graphql": "^14.0.0",
     "graphql-tag": "^2.9.2",
-    "moment": "^2.19.2",
-    "react": "^16.4.0",
-    "react-apollo": "^2.1.9",
-    "react-dom": "^16.4.0",
-    "react-markdown": "^3.3.4",
-    "react-router-dom": "^4.2.2",
-    "react-swipeable-views": "^0.12.14",
-    "styled-components": "3.3.3",
+    "moment": "^2.22.2",
+    "react": "^16.4.2",
+    "react-apollo": "^2.1.11",
+    "react-dom": "^16.4.2",
+    "react-loadable": "^5.5.0",
+    "react-markdown": "^3.5.0",
+    "react-router-dom": "^4.3.1",
+    "react-swipeable-views": "^0.12.17",
+    "styled-components": "3.4.5",
     "uuid": "^3.3.2"
   },
   "devDependencies": {
-    "@babel/core": "v7.0.0-beta.53",
-    "@babel/node": "7.0.0-beta.53",
-    "@babel/preset-env": "v7.0.0-beta.53",
-    "@babel/preset-react": "v7.0.0-beta.53",
-    "@babel/preset-stage-3": "^7.0.0-beta.53",
-    "@babel/preset-typescript": "v7.0.0-beta.53",
-    "@types/react": "^16.4.6",
-    "@types/react-dom": "^16.0.5",
-    "@types/react-router-dom": "^4.2.6",
+    "@babel/core": "^7.0.0",
+    "@babel/node": "^7.0.0",
+    "@babel/plugin-proposal-class-properties": "^7.0.0",
+    "@babel/plugin-proposal-json-strings": "^7.0.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
+    "@babel/plugin-syntax-import-meta": "^7.0.0",
+    "@babel/preset-env": "^7.0.0",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/preset-typescript": "v7.0.0",
+    "@types/react": "^16.4.13",
+    "@types/react-dom": "^16.0.7",
+    "@types/react-loadable": "^5.4.1",
+    "@types/react-router-dom": "^4.3.0",
     "@types/react-swipeable-views": "^0.12.1",
-    "@types/uuid": "^3.4.3",
-    "babel-loader": "^8.0.0-beta.4",
+    "@types/uuid": "^3.4.4",
+    "babel-loader": "^8.0.2",
+    "babel-plugin-import": "^1.8.0",
+    "babel-plugin-styled-components": "^1.6.0",
     "connect-history-api-fallback": "^1.5.0",
     "koa-connect": "^2.0.1",
-    "ts-loader": "^4.4.2",
-    "tslint": "^5.10.0",
-    "tslint-config-airbnb": "^5.9.2",
+    "ts-loader": "^5.0.0",
+    "tslint": "^5.11.0",
+    "tslint-config-airbnb": "^5.11.0",
     "tslint-loader": "^3.6.0",
     "tslint-react": "^3.6.0",
-    "typedoc": "^0.11.1",
-    "typescript": "^2.9.2",
+    "typedoc": "^0.12.0",
+    "typescript": "^3.0.3",
     "typescript-styled-plugin": "^0.10.0",
-    "webpack": "^4.16.0",
-    "webpack-cli": "^3.0.8",
+    "uglifyjs-webpack-plugin": "^1.3.0",
+    "webpack": "^4.17.1",
+    "webpack-bundle-analyzer": "^2.13.1",
+    "webpack-bundle-size-analyzer": "^2.7.0",
+    "webpack-cli": "^3.1.0",
     "webpack-serve": "^2.0.2"
   }
 }
