@@ -1,14 +1,16 @@
-const { DefinePlugin } = require ("webpack")
-const path             = require("path");
-const convert          = require('koa-connect');
-const history          = require('connect-history-api-fallback');
+const { DefinePlugin }     = require ("webpack")
+const path                 = require("path");
+const convert              = require('koa-connect');
+const history              = require('connect-history-api-fallback');
+const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin
 
 module.exports = {
     entry: "./src/index.tsx",
     mode: "development",
     output: {
         filename: "[name].js",
-        path: path.resolve(__dirname, "dist")
+        path: path.resolve(__dirname, "dist"),
+        publicPath: "/"
     },
     resolve: {
         modules: ["node_modules"],
@@ -28,6 +30,7 @@ module.exports = {
         ]
     },
     plugins: [
+        // new BundleAnalyzerPlugin(),
         new DefinePlugin(
             Object.entries(process.env)
                 .map(x => ({["process.env." + x[0]]: JSON.stringify(x[1])}))
