@@ -1,12 +1,17 @@
-const { DefinePlugin } = require ("webpack")
-const path             = require("path");
+const { DefinePlugin }     = require ("webpack")
+const path                 = require("path");
+const convert              = require('koa-connect');
+const history              = require('connect-history-api-fallback');
+const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
+const Uglify               = require("uglifyjs-webpack-plugin");
 
 module.exports = {
     entry: "./src/index.tsx",
     mode: "production",
     output: {
         filename: "[name].js",
-        path: path.resolve(__dirname, "dist")
+        path: path.resolve(__dirname, "dist"),
+        publicPath: "/"
     },
     resolve: {
         modules: ["node_modules"],
@@ -26,10 +31,31 @@ module.exports = {
         ]
     },
     plugins: [
+        // new Uglify({
+        //     sourceMap: false,
+        //     uglifyOptions: {
+        //         ecma: 8,
+        //         // mangle: {
+        //         //     properties: true
+        //         // },
+        //         output: {
+        //             // comments: "@license",
+        //             beautify: false
+        //         },
+        //     }
+        // }),
+        // new BundleAnalyzerPlugin(),
         new DefinePlugin(
             Object.entries(process.env)
                 .map(x => ({["process.env." + x[0]]: JSON.stringify(x[1])}))
                 .reduce((x, y) => Object.assign(x, y), {}),
         )
-    ]
+    ],
+    serve: {
+        add: (app, middleware, options) => {
+            app.use(convert(history()));
+        },
+        content: path.resolve(__dirname, 'assets'),
+        open: true
+    }
 };
